Change the way comments are formatted, displaying first name and last initial, along with the email associated with the sign in. Delete all button, even if called by a user, will only execute the deletes if the user is an admin

+++ portfolio/src/main/java/com/google/sps/servlets/deleteAll.java
@@ -43,7 +43,7 @@ public class deleteAll extends HttpServlet{
         Query query = new Query("Comment");
         DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
         PreparedQuery results = datastore.prepare(query);
-        if (userService.isUserAdmin()) {
+        if (userService.isUserLoggedIn() && userService.isUserAdmin()) {
             for (Entity entity : results.asIterable()) {
             Key key = entity.getKey();
             datastore.delete(key);
